Crate heim_disk

Disks information.

Modules

os
OS-specific extensions.

Structs

IoCounters
Disk I/O counters.

Partition
Mounted disk partition.

Usage
Disk usage statistics.

Enums

FileSystem
Known filesystems.

Functions

io_counters
Returns a stream over IO counters for all disks available in system.

io_counters_physical
Returns a stream over IO counters for each physical disk installed on the system.

partitions
Returns a stream over mounted disk Partitions.

partitions_physical
Returns a stream over physical only mounted disk Partitions.

usage
Returns disk Usage statistics about the partition which contains the given path.